Organização de código app compilacao
- Organização de Códigos duplicados na app compilação
- Criação da common_tags na app base
