+++ base/templatetags/common_tags.py
@@ -1,0 +1,39 @@
+from compressor.utils import get_class
+from django import template
+
+register = template.Library()
+
+
+@register.simple_tag
+def field_verbose_name(instance, field_name):
+    return instance._meta.get_field(field_name).verbose_name
+
+
+@register.simple_tag
+def fieldclass_verbose_name(class_name, field_name):
+    cls = get_class(class_name)
+    return cls._meta.get_field(field_name).verbose_name
+
+
+@register.simple_tag
+def model_verbose_name(class_name):
+    model = get_class(class_name)
+    return model._meta.verbose_name
+
+
+@register.simple_tag
+def model_verbose_name_plural(class_name):
+    model = get_class(class_name)
+    return model._meta.verbose_name_plural
+
+
+@register.filter
+def lookup(d, key):
+    skey = str(key)
+    return d[str(key)] if skey in d else []
+
+
+@register.filter
+def isinst(value, class_str):
+    classe = value.__class__.__name__
+    return classe == class_str
