+++ compilacao/templatetags/compilacao_filters.py
@@ -1,4 +1,3 @@
-from compressor.utils import get_class
 from django import template
 from django.core.signing import Signer
 from django.db.models import Q
@@ -20,12 +19,6 @@ def get_tipos_dispositivo(pk_atual):
 
     return TipoDispositivo.objects.filter(
         id__gte=pk_atual)
-
-
-@register.filter
-def lookup(d, key):
-    skey = str(key)
-    return d[str(key)] if skey in d else []
 
 
 @register.simple_tag
@@ -153,31 +146,6 @@ def nomenclatura_heranca(d, ignore_ultimo=0, ignore_primeiro=0):
         d = d.dispositivo_pai
 
     return result
-
-
-@register.simple_tag
-def field_verbose_name(instance, field_name):
-    return instance._meta.get_field(field_name).verbose_name
-
-
-@register.simple_tag
-def fieldclass_verbose_name(class_name, field_name):
-    cls = get_class(
-        'compilacao.models.' + class_name)
-    return cls._meta.get_field(
-        field_name).verbose_name
-
-
-@register.simple_tag
-def model_verbose_name(class_name):
-    model = get_class('compilacao.models.' + class_name)
-    return model._meta.verbose_name
-
-
-@register.simple_tag
-def model_verbose_name_plural(class_name):
-    model = get_class('compilacao.models.' + class_name)
-    return model._meta.verbose_name_plural
 
 
 @register.filter
