--- compilacao/views2.py
@@ -1,54 +1,0 @@
-from django.utils.translation import ugettext_lazy as _
-
-from compilacao.models import PerfilEstruturalTextoArticulado, TipoDispositivo
-from sapl.crud import build_crud
-
-perfil_estr_txt_norm = build_crud(
-    PerfilEstruturalTextoArticulado, 'perfil_estrutural', [
-
-        [_('Perfil Estrutural de Textos Articulados'),
-         [('sigla', 2), ('nome', 10)]],
-    ])
-
-
-tipo_dispositivo_crud = build_crud(
-    TipoDispositivo, 'tipo_dispositivo', [
-
-        [_('Dados Básicos'),
-         [('nome', 8), ('class_css', 4)]],
-
-        [_('Configurações para Edição do Rótulo'),
-         [('rotulo_prefixo_texto', 3),
-          ('rotulo_sufixo_texto', 3),
-          ('rotulo_ordinal', 3),
-          ('contagem_continua', 3)],
-
-         ],
-
-        [_('Configurações para Renderização de Rótulo e Texto'),
-         [('rotulo_prefixo_html', 6),
-          ('rotulo_sufixo_html', 6), ],
-
-         [('texto_prefixo_html', 4),
-          ('dispositivo_de_articulacao', 4),
-          ('texto_sufixo_html', 4)],
-         ],
-
-        [_('Configurações para Nota Automática'),
-         [('nota_automatica_prefixo_html', 6),
-          ('nota_automatica_sufixo_html', 6),
-          ],
-         ],
-
-        [_('Configurações para Variações Numéricas'),
-
-         [('formato_variacao0', 12)],
-         [('rotulo_separador_variacao01', 5), ('formato_variacao1', 7), ],
-         [('rotulo_separador_variacao12', 5), ('formato_variacao2', 7), ],
-         [('rotulo_separador_variacao23', 5), ('formato_variacao3', 7), ],
-         [('rotulo_separador_variacao34', 5), ('formato_variacao4', 7), ],
-         [('rotulo_separador_variacao45', 5), ('formato_variacao5', 7), ],
-
-         ],
-
-    ])
