--- templates/compilacao/detail.html
@@ -1,38 +1,0 @@
-{% extends "base.html" %}
-{% load i18n %}
-
-{% block base_content %}
-
-  {# FIXME is this the best markup to use? #}
-  <div class="clearfix">
-    {% block actions %}
-      <dl class="sub-nav right">
-        <dd><a href="{{ view.update_url }}" class="button">{% trans 'Editar' %}</a></dd>
-        <dd><a href="{{ view.delete_url }}" class="button alert">{% trans 'Excluir' %}</a></dd>
-      </dl>
-    {% endblock actions %}
-    {% block sections_nav %}{% endblock %}
-  </div>
-
-  {% block detail_content %}
-    {# TODO replace fieldset for something semantically correct, but with similar visual grouping style #}
-    {% for fieldset in view.fieldsets %}
-    <fieldset>
-      <legend>{{ fieldset.legend }}</legend>
-      {% for row in fieldset.rows %}
-      <div class="row">
-          {% for column in row %}
-          <div class="columns large-{{ column.span }}">
-              <div id="div_id_{{ column.id }}" class="holder">
-                  <label>{{ column.verbose_name }}</label> {# TODO replace labels, probably (are they correct here?) #}
-                  <p>{{ column.text }}</p>
-              </div>
-          </div>
-          {% endfor %}
-      </div>
-      {% endfor %}
-    </fieldset>
-    {% endfor %}
-  {% endblock detail_content %}
-
-{% endblock base_content %}
