+++ templates/compilacao/text_edit.html
@@ -4,6 +4,7 @@
 {% load staticfiles %}
 {% load sass_tags %}
 {% load crispy_forms_tags %}
+{% load common_tags %}
 
 
 {% block head_content %}{{block.super}}
