+++ templates/compilacao/text_edit_bloco.html
@@ -1,5 +1,6 @@
 {% load i18n %}
 {% load compilacao_filters %}
+{% load common_tags %}
 
 {% for dpt in object_list %}
 	{% if dpt.nivel == view.flag_nivel_old %}
