+++ templates/compilacao/text_list.html
@@ -1,6 +1,7 @@
 {% extends "compilacao/textoarticulado_detail.html" %}
 {% load i18n %}
 {% load compilacao_filters %}
+{% load common_tags %}
 {% load staticfiles %}
 {% load sass_tags %}
