+++ templates/compilacao/text_list_bloco.html
@@ -1,5 +1,6 @@
 {% load i18n %}
 {% load compilacao_filters %}
+{% load common_tags %}
 
 <div class="display"></div>
