+++ templates/compilacao/text_list_blocoalteracao.html
@@ -1,4 +1,5 @@
 {% load compilacao_filters %}
+{% load common_tags %}
 <i class="fa fa-quote-left fa-2x fa-pull-left fa-border"></i>
 {% for ch in dpt.pk|get_bloco_atualizador %}
 {% spaceless %}
