+++ templates/compilacao/textoarticulado_list.html
@@ -1,26 +1,17 @@
-{% extends "base.html" %}
+{% extends "compilacao/textoarticulado_detail.html" %}
 {% load i18n %}
 {% load compilacao_filters %}
+{% load common_tags %}
 
 {% block base_content %}
-
-
-{# FIXME is this the best markup to use? #}
-
-   <a href="#" class="button tiny right" data-dropdown="dropconfig"><i class="fi-widget"></i></a>
-      <ul id="dropconfig" class="tiny f-dropdown" data-dropdown-content>
-          <li><a href="{% url 'tipo_ta_list' %}">{%model_verbose_name_plural 'TipoTextoArticulado'%}</a></li>
-          <li><a href="{% url 'tipopublicacao:list' %}">{%model_verbose_name_plural 'TipoPublicacao'%}</a></li>
-          <li><a href="{% url 'veiculopublicacao:list' %}">{%model_verbose_name_plural 'VeiculoPublicacao'%}</a></li>
-          <li><a href="{% url 'tiponota:list' %}">{%model_verbose_name_plural 'TipoNota'%}</a></li>
-          <li><a href="{% url 'tipovide:list' %}">{%model_verbose_name_plural 'TipoVide'%}</a></li>
-          <li><a href="#">Tipo de Dispositivo</a></li>
-          <li><a href="#">Perfil Estrutural de Textos Articulados</a></li>
-      </ul>
-
+    {% block actions %}{{block.super}}
+    
+    {% endblock actions %}
+{% block detail_content %} 
+{% endblock detail_content %} 
 <dl class="sub-nav">
   <dd><a href="{{ view.create_url }}">
-    {% trans 'Adicionar'%} {%model_verbose_name 'TextoArticulado'%}
+    {% trans 'Adicionar'%} {%model_verbose_name 'compilacao.models.TextoArticulado'%}
 
   </a></dd>
 
@@ -33,11 +24,11 @@
 <table class="table table-hover">
     <thead>
         <tr>
-            <th>{% trans 'Tipo'%}</th>
-            <th>{% trans 'Número'%}</th>
-            <th>{% trans 'Ano'%}</th>
-            <th>{% trans 'Data'%}</th>
-            <th>{% trans 'Ementa'%}</th>
+            <th>{% fieldclass_verbose_name 'compilacao.models.TextoArticulado' 'tipo_ta' %}</th>
+            <th>{% fieldclass_verbose_name 'compilacao.models.TextoArticulado' 'numero' %}</th>
+            <th>{% fieldclass_verbose_name 'compilacao.models.TextoArticulado' 'ano' %}</th>
+            <th>{% fieldclass_verbose_name 'compilacao.models.TextoArticulado' 'data' %}</th>
+            <th>{% fieldclass_verbose_name 'compilacao.models.TextoArticulado' 'ementa' %}</th>
         </tr>
     </thead>
     <tbody>
@@ -52,40 +43,5 @@
         {% endfor %}
     </tbody>
 </table>
-
-{% if is_paginated %}
-<div class="pagination-centered">
-  <ul class="pagination" role="menubar" aria-label="Pagination">
-    {% if page_obj.has_previous %}
-    <li>
-      <a href="?page={{ page_obj.previous_page_number }}">
-        <span class="arrow">&laquo;</span>
-      </a>
-    </li>
-    {% else %}
-    <li class="arrow unavailable" aria-disabled="true"><a href="">&laquo;</a></li>
-    {% endif %}
-
-    {% for page in page_range %}
-        {% if page %}
-            <li {% if page == page_obj.number %}class="current"{% endif %}>
-              <a href="?page={{ page }}">{{ page }}</a>
-            </li>
-            {% else %}
-            <li class="unavailable" aria-disabled="true"><a href="">&hellip;</a></li>
-        {% endif %}
-    {% endfor %}
-
-    {% if page_obj.has_next %}
-    <li>
-      <a href="?page={{ page_obj.next_page_number }}">
-        <span class="arrow">&raquo;</span>
-      </a>
-    </li>
-    {% else %}
-    <li class="arrow unavailable" aria-disabled="true"><a href="">&raquo;</a></li>
-    {% endif %}
-  </ul>
-</div>
-{% endif %}
+{% include 'paginacao.html'%}
 {% endblock %}
