+++ templates/compilacao/tipotextoarticulado_detail.html
@@ -1,4 +1,5 @@
 {% extends "base.html" %} {% load i18n %} {% load compilacao_filters %}
+{% load common_tags %}
 
 {% block base_content %} {# FIXME is this the best markup to use? #}
